+++ lib/tynn/settings.rb
@@ -1,5 +1,5 @@
 class Tynn
-  module Settings # :nodoc: all
+  module Settings # :nodoc:
     def self.deepclone(hash)
       default_proc, hash.default_proc = hash.default_proc, nil
 
@@ -8,13 +8,7 @@ class Tynn
       hash.default_proc = default_proc
     end
 
-    module InstanceMethods # :nodoc: all
-      def settings
-        return self.class.settings
-      end
-    end
-
-    module ClassMethods # :nodoc: all
+    module ClassMethods # :nodoc:
       def inherited(subclass)
         subclass.settings.replace(Tynn::Settings.deepclone(settings))
         subclass.settings.default_proc = proc { |h, k| h[k] = settings[k] }